+++ src/modules/profile.ts
@@ -1,6 +1,11 @@
-import { HookFunction } from "../util/sdk";
+import { FilterArrayFromArray } from "../util/general";
+import { bcxAPI, bcxFound, HookFunction } from "../util/sdk";
 import { Module, ModuleTitle } from "./_module";
-// import { UpdateElementValues } from "./settings";
+import { ElementName } from "./settings";
+import { ApplyPetHearing, NonDisruptivePetSpeech } from "../util/petHearAndSpeak";
+import { HookedMessage, MPAMessageContent, SendMPAMessage } from "../util/messaging";
+import { RecordsSync } from "./dataSync";
+import { IsMemberNumberInAuthGroup } from "../util/authority";
 
 const PlayerP: (C?: Character) => MPARecord = (C: Character = Player) =>
 {
@@ -18,18 +23,48 @@ function GetRecord(title: ModuleTitle, C: Character = Player): MPARecord
 }
 
 // Make sure not in BCX banned words / phrases
+// Don't blame me for the selection, how are you supposed to type out these animal sounds
+// Half this came from chat gpt anyway, if you got something better, feel free to suggest it
 export const GARBLE_PHRASES = Object.freeze({
-    Human: "Hmmgm",
-    Custom: "",
-    Bunny: "Pon",
-    Cat: "Mew",
-    Cow: "Moo",
-    Dog: "Arf",
-    Fox: "Chirp",
-    Mouse: "Sqeak",
-    Pony: "Neigh",
-    Wolf: "Grrr"
+    Human: ["hmmgm", "mmmm", "mhhmmaa", "hmmm"],
+    Custom: [""],
+    Bunny: ["pon", "eep", "huff", "tk-tk", "nngh-nngh"],
+    Cat: ["meow", "mew", "nyah", "nya", "purr", "mrrow", "prrr"],
+    Cow: ["moo", "mooooo", "muuhhh", "moooah"],
+    Dog: ["arf", "woof", "bark", "ruff", "huff", "eep"],
+    Fox: ["chirp", "yip", "wow-wow", "yoww", "eep"],
+    Mouse: ["squeak", "eeek", "tch-tch", "wi-i-i", "sss"],
+    Pony: ["neigh", "whinny", "wheee", "snrt", "whooosh", "huff", "nnhhh-ker"],
+    Wolf: ["awoo", "aa-ooo", "woof", "ruff", "arf", "wiiip", "huff"]
 });
+type PetSpeakingKeys = keyof typeof GARBLE_PHRASES;
+
+/**
+ * Get the pet garble phrases for the selected profile
+ * @param profile - Defaults to Player's profile if none provided
+ */
+function GetPetGarblePhrases(profile: PetSpeakingKeys = PlayerP(Player).type): string[]
+{
+    if (profile in GARBLE_PHRASES && profile !== "Custom")
+    {
+        return GARBLE_PHRASES[profile];
+    }
+    return (PlayerP(Player).garblePhrases as string).split(",").map((str) => str.trim());
+}
+
+/**
+ * Only include pet sounds that are permitted by BCX
+ * @param profile - Defaults to Player's profile
+ * @returns Phrases BCX won't get mad at
+ */
+function GetAllowedPetGarblePhrases(profile: PetSpeakingKeys = PlayerP(Player).type): string[]
+{
+    const bannedBCXWords = bcxAPI()?.getRuleState("speech_ban_words")?.customData?.bannedWords ?? [];
+    const petPhrases = bcxFound() && PlayerP(Player).bcxSpeaking ?
+        bcxAPI()?.getRuleState("speech_mandatory_words")?.customData?.mandatoryWords ?? [] :
+        GetPetGarblePhrases(profile);
+    return FilterArrayFromArray(petPhrases, bannedBCXWords);
+}
 
 export const PET_HEARING = Object.freeze({
     All: ["owner", "pet", "master", "mistress", "miss", "sir", "paw", "sit", "come", "down", "up",
@@ -48,12 +83,119 @@ export const PET_HEARING = Object.freeze({
         "stomp", "calm", "easy", "slow", "cart", "bay", "run", "race"],
     Wolf: ["bone", "wolf", "dog", "pup", "puppy", "heel"]
 });
+type PetHearingKeys = keyof typeof PET_HEARING;
+
+/**
+ *
+ * @param profile - Defaults to Player's profile
+ */
+function GetPetHearingPhrases(profile: PetHearingKeys = PlayerP(Player).type): string[]
+{
+    if (profile in PET_HEARING)
+    {
+        return PET_HEARING.All.concat(PET_HEARING[profile]);
+    }
+
+    return PET_HEARING.All;
+}
+
+/**
+ * Callback when profile type is set
+ * @param C - Character who we are setting
+ * @param value - What the new type is
+ * @param prevValue - What the previous type was
+ */
+function SetProfileType(C: Character, value: string, prevValue: string)
+{
+    const eleID = ElementName(ModuleTitle.Profile, "garblePhrases");
+    if (value === "Custom")
+    {
+        PlayerP(C).garblePhrases = PlayerP(C).customGarblePhrases.join(", ");
+        if (document.getElementById(eleID))
+        {
+            ElementValue(eleID, PlayerP(C).garblePhrases);
+        }
+        return;
+    }
+    if (prevValue === "Custom" && document.getElementById(eleID))
+    {
+        PlayerP(C).customGarblePhrases = ElementValue(eleID).split(",").map((x) => x.trim());
+    }
+    PlayerP(C).garblePhrases = (GARBLE_PHRASES[PlayerP(C).type] as string[]).join(", ");
+
+    if (document.getElementById(eleID))
+    {
+        ElementValue(eleID, PlayerP(C).garblePhrases);
+    }
+}
+
+/**
+ * Set profile bcx speaking for the local player
+ */
+function SetBCXSpeak()
+{
+    PlayerP().bcxSpeaking = true;
+    const eleID = ElementName(ModuleTitle.Profile, "garblePhrases");
+    // Save previous words if needed
+    if (PlayerP().type === "Custom" && document.getElementById(eleID))
+    {
+        PlayerP().customGarblePhrases = ElementValue(eleID).split(",").map((x) => x.trim());
+    }
+    const words = bcxAPI()?.getRuleState("speech_mandatory_words")?.customData?.mandatoryWords ?? [];
+    PlayerP().garblePhrases = words.join(", ");
+    if (document.getElementById(eleID))
+    {
+        ElementValue(eleID, PlayerP().garblePhrases);
+    }
+}
 
 export class ProfileModule extends Module
 {
     get Title(): ModuleTitle
     {
         return ModuleTitle.Profile;
+    }
+
+    get SyncListeners(): HookedMessage[]
+    {
+        return [
+            {
+                module: this.Title,
+                message: "BCXSpeakEnableRequest",
+                action: function (sender: Character, _content: MPAMessageContent): void
+                {
+                    // Check if the sender has authority to set this setting
+                    if (!IsMemberNumberInAuthGroup(sender.MemberNumber ?? -1, Player.MPA?.[ModuleTitle.Authority][`others${ModuleTitle.Profile}`]))
+                    {
+                        console.warn(`${sender.Nickname || sender.Name}[${sender.MemberNumber}] tried to illegially set your bcx speak to be enabled`);
+                        return;
+                    }
+
+                    // Set the record by the person who is having their settings changed
+                    // BCX mandatory words only viewable by the player locally
+                    SetBCXSpeak();
+                    RecordsSync([
+                        { category: ModuleTitle.Profile, record: "garblePhrases" },
+                        { category: ModuleTitle.Profile, record: "bcxSpeaking" }
+                    ], sender.MemberNumber);
+
+                    // After updating and sending the new records, notify sender for acknowledgement to update their display
+                    SendMPAMessage({ message: "BCXSpeakEnableReply" }, sender.MemberNumber);
+                }
+            }, {
+                module: this.Title,
+                message: "BCXSpeakEnableReply",
+                action: function (sender: Character, _content: MPAMessageContent): void
+                {
+                    // We got a postive response from the request, change made, but we need to see it on our side now
+                    const eleID = ElementName(ModuleTitle.Profile, "garblePhrases");
+                    if (document.getElementById(eleID))
+                    {
+                        ElementValue(eleID, PlayerP(sender).garblePhrases);
+                    }
+                }
+            }
+        ];
     }
 
     get Settings(): Setting[]
@@ -66,27 +208,71 @@ export class ProfileModule extends Module
                 options: Object.keys(GARBLE_PHRASES),
                 value: "Human",
                 label: "Use a preset profile or make your own",
-                loop: true
-                // onSet(C)
-                // {
-                //     PlayerP(C).garblePhrases = GARBLE_PHRASES[PlayerP(C).type];
-                //     UpdateElementValues();
-                // }
+                loop: true,
+                onSet: (C, value, prevValue) =>
+                {
+                    if (PlayerP(C).bcxSpeaking)
+                    {
+                        return;
+                    }
+                    SetProfileType(C, value, prevValue);
+                }
             } as OptionSetting, {
                 name: "petHearing",
                 type: "checkbox",
                 active: () => true,
                 value: false,
                 label: "While deafened certain pet keywords can be picked up"
-            } as CheckboxSetting, /* {
+            } as CheckboxSetting, {
                 name: "garblePhrases",
                 type: "text",
-                active: (C) => PlayerP(C).type === "Custom",
-                value: GARBLE_PHRASES.Human,
-                label: "Phrases you speak when unable to be understood",
+                active: (C) => PlayerP(C).type === "Custom" && !(PlayerP(C).petSpeaking && bcxFound() && PlayerP(C).bcxSpeaking),
+                value: GARBLE_PHRASES.Human.join(),
+                label: "Pet sounds you make",
                 maxChars: 1024,
-                width: 400
-            } as TextSetting, */{
+                width: 1024
+            } as TextSetting, {
+                name: "customGarblePhrases",
+                type: "record",
+                value: []
+            } as Setting, {
+                name: "petSpeaking",
+                type: "checkbox",
+                active: () => true,
+                value: false,
+                label: "Automatically make the above pet sounds while talking"
+            } as CheckboxSetting, {
+                name: "bcxSpeaking",
+                type: "checkbox",
+                active: (C) => PlayerP(C).petSpeaking && bcxFound(),
+                value: false,
+                label: "Use BCX mandatory words instead of the pet sounds above",
+                onSet: (C, value) =>
+                {
+                    if (!value)
+                    {
+                        SetProfileType(C, PlayerP(C).type, "");
+                        return;
+                    }
+                    if (C.IsPlayer())
+                    {
+                        // Update the mandatory words for yourself since have local bcx access
+                        SetBCXSpeak();
+                    }
+                    else
+                    {
+                        const eleID = ElementName(ModuleTitle.Profile, "garblePhrases");
+                        // Save any custom word changes
+                        if (PlayerP(C).type === "Custom" && document.getElementById(eleID))
+                        {
+                            PlayerP(C).customGarblePhrases = ElementValue(eleID).split(",").map((x) => x.trim());
+                        }
+
+                        // Request the mandatory words from the other person
+                        SendMPAMessage({ message: "BCXSpeakEnableRequest" }, C.MemberNumber);
+                    }
+                }
+            } as CheckboxSetting, {
                 name: "hardcore",
                 type: "checkbox",
                 active: () => true,
@@ -114,7 +300,9 @@ export class ProfileModule extends Module
     Load(): void
     {
         super.Load();
+        SpeechTransformAllEffects.push("petSpeak" as SpeechTransformName);
 
+        // Pet hearing
         HookFunction(this.Title, "SpeechTransformGagGarble", 5, ([text, intensity, ignoreOOC, ...args], next) =>
         {
             // MBS adds Character parameter to this, can use as an extra check for when to apply
@@ -128,74 +316,33 @@ export class ProfileModule extends Module
                 || PlayerP().petHearing === false
             )
             {
-                return next([text, intensity, ignoreOOC, ...args]);
+                return next([text, intensity, ignoreOOC]);
             }
-
-            // Split the input text into different substrings exluding the words you want to keep
-            // Step 1: Get the words you want to keep if any
-            const phrasesToKeep = PET_HEARING.All.concat(PET_HEARING[PlayerP()?.type]);
-            // Add the Player's name as a word they can hear
-            phrasesToKeep.push((Player.Nickname || Player.Name).toLocaleLowerCase());
-            // Add the Player's gender as in the example, "Good girl"
-            phrasesToKeep.push(Player.GetPronouns() === "SheHer" ? "girl" : "boy");
-
-            // Step 2: Find the matches if they exist in the orginal message
-            // Check lower case cuz the phrase we looking for in lower case, does not matter for final result, only want position
-            const message = text.toLocaleLowerCase();
-            const foundPhrases: { start: number; len: number }[] = [];
-            phrasesToKeep.forEach((phrase) =>
-            {
-                const regExp = new RegExp(`\\b${phrase}\\b`, "g");
-                const regexMatches = [...message.matchAll(regExp)];
-                regexMatches.forEach((match) =>
-                {
-                    foundPhrases.push({ start: match.index, len: phrase.length });
-                });
-            });
-
-            // No matches found so can skip and early return
-            if (foundPhrases.length === 0)
-            {
-                return next([text, intensity, ignoreOOC, ...args]);
-            }
-
-            // Sort so that it is constructed in the correct order when put back together; lower starting indexes first
-            foundPhrases.sort((a, b) => a.start - b.start);
-            // Remove any phrases that are found within OOC speech range;
-            const oocRanges = SpeechGetOOCRanges(text);
-            oocRanges.forEach((range) =>
-            {
-                for (let i = foundPhrases.length - 1; i >= 0; i--)
-                {
-                    const phrase = foundPhrases[i];
-                    // Remove if the phrase starts or ends within the ooc range
-                    if (
-                        (range.start < phrase.start && phrase.start < range.start + range.length)
-                        || (range.start < phrase.start + phrase.len && phrase.start + phrase.len < range.start + range.length)
-                    )
-                    {
-                        foundPhrases.splice(i, 1);
-                    }
-                }
-            });
-
-            // May need to check that one phrase does not start before another ends, and if so merge into a single longer entry
-
-            let lastIndex = 0;
-            let finalText = "";
-            foundPhrases.forEach((found) =>
-            {
-                // Before the phrase to omit
-                finalText += next([text.substring(lastIndex, found.start), intensity, ignoreOOC, ...args]);
-
-                // Phrase we want to omit
-                lastIndex = found.start + found.len;
-                finalText += text.substring(found.start, found.start + found.len);
-            });
-            // Last of the text to garble up
-            finalText += next([text.substring(lastIndex), intensity, ignoreOOC, ...args]);
-
-            return finalText;
+            return ApplyPetHearing([text, intensity, ignoreOOC], next, GetPetHearingPhrases());
         });
+
+        // Insert pet sounds on input for BCX support
+        HookFunction(this.Title, "CommandParse", 8, ([msg], next) =>
+        {
+            // Skip commands
+            // Skip if not attempting to apply pet speech
+            msg = msg.trim();
+            if (
+                [CommandsKey, "*", "@"].includes(msg.charAt(0))
+                || !PlayerP(Player).petSpeaking
+            )
+            {
+                return next([msg]);
+            }
+
+            return NonDisruptivePetSpeech(msg, next, GetAllowedPetGarblePhrases());
+        });
+    }
+
+    Unload(): void
+    {
+        super.Unload();
+
+        SpeechTransformAllEffects = SpeechTransformAllEffects.filter((transform) => transform as SpeechTransformName | "petSpeak" !== "petSpeak");
     }
 }