+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.4.4
+ - Pet Hearing bug fix, no more duplicate hearing
+ - Adding pet sounds to input speech; BCX compatible
+ - Lockout player out of settings options
+ - Pet sounds can use BCX mandatory words instead
+ - Notification when settings are being accessed by others
+
 # 0.4.3
  - Pet hearing based on profile
  - Wiki and link to wiki from MPA settings page